CHUKWA-121.  Added logic to detect partition number less than 0.  (Eric Yang)
git-svn-id: https://svn.apache.org/repos/asf/hadoop/chukwa/trunk@765369 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/hadoop/chukwa/database/DataExpiration.java
@@ -68,9 +68,11 @@ public class DataExpiration {
                 + partition + "_" + parts[parts.length - 1];
             dbw.execute(dropPartition);
             partition--;
-            dropPartition = "drop table if exists " + table + "_" + partition
-                + "_" + parts[parts.length - 1];
-            dbw.execute(dropPartition);
+            if(partition>=0) {
+              dropPartition = "drop table if exists " + table + "_" + partition
+                  + "_" + parts[parts.length - 1];
+              dbw.execute(dropPartition);
+            }
           } catch (NumberFormatException e) {
             log
                 .error("Error in parsing table partition number, skipping table:"
@@ -81,6 +83,7 @@ public class DataExpiration {
           }
         }
       }
+      dbw.close();
     } catch (Exception e) {
       e.printStackTrace();
     }
